Move plugin manager model into labext.
