+++ jupyterlab/myplugins/fooplugin.py
@@ -1,4 +1,4 @@
-from flexx_jlab_serverext import register_flexx_jlab_plugin
+from flexx_labext import register_flexx_jlab_plugin
 
 from flexx import app, event, ui
